рефакторинг: создан файл utils.js в него вынесены некоторые общие функции; удаление лишних приведений типов в trip-point.js
