+++ src/mock/trip-point.js
@@ -1,15 +1,7 @@
 import dayjs from 'dayjs';
 import {generateDestinations} from './destinations.js';
 
-// from you dont need lodash underscore
-export const randomInt = (a = 1, b = 0) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-  return Math.floor(lower + Math.random() * (upper - lower + 1));
-};
-
-export const getRandomArrayElement = (arr) => arr[randomInt(0, arr.length - 1)];
-
+import {randomInt, getRandomArrayElement} from '../utils.js';
 
 const OFFERS = {
   'taxi': [
