+++ src/view/edit-trip-point.js
@@ -1,12 +1,7 @@
 import dayjs from 'dayjs';
 import {POINT_TYPES} from '../mock/trip-point.js';
 
-// from you dont need lodash underscore
-export const randomInt = (a = 1, b = 0) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-  return Math.floor(lower + Math.random() * (upper - lower + 1));
-};
+import {randomInt} from '../utils.js';
 
 export const getEditPointTemplate = (point = {}, destinationsAvailable = []) => {
   const isEmpty = Object.keys(point).length === 0;
